Ball object now works with sprite class. Needs refactoring and adding to all game objects

+++ Breakout/source/Text_Renderer.cpp
@@ -80,7 +80,7 @@ void Text_Renderer::Load(std::string font, GLuint font_size)
 			texture,
 			glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x
+			static_cast<GLuint>(face->glyph->advance.x)
 		};
 		_characters.insert(std::pair<GLchar, Character>(c, character));
 	}
